Uncommitted active diff selection
I don't love this but it feels like a lesser evil vs hoisting all diff view production, or somehow syncing them with refs (given async).
