+++ apps/lite/ui/src/routes/project/$id/workspace/Details.tsx
@@ -33,12 +33,10 @@ import {
 	type FileOperand,
 	fileOperand,
 	hunkOperand,
-	operandIdentityKey,
 	type FileParent,
 	type HunkOperand,
 	type Operand,
 	weakCommitIdentityKey,
-	weakFileIdentityKey,
 } from "#ui/operands.ts";
 import type { BranchTab } from "#ui/projects/project.ts";
 import { projectSlice } from "#ui/projects/state.ts";
@@ -58,20 +56,13 @@ import {
 	FieldTextareaStyles,
 } from "#ui/components/Field.tsx";
 import { Field, Toggle, ToggleGroup, Toolbar, Tooltip } from "@base-ui/react";
+import type { CiCheck, CommitDetails, TreeChange } from "@gitbutler/but-sdk";
 import type {
-	CiCheck,
-	CommitDetails,
-	DiffHunk,
-	TreeChange,
-	UnifiedPatch,
-} from "@gitbutler/but-sdk";
-import {
-	type CodeViewDiffItem,
-	type CodeView as CodeViewClass,
-	type CodeViewLineSelection,
-	parsePatchFiles,
-	type GetHoveredLineResult,
-	type DiffLineAnnotation,
+	CodeViewDiffItem,
+	CodeView as CodeViewClass,
+	CodeViewLineSelection,
+	GetHoveredLineResult,
+	DiffLineAnnotation,
 } from "@pierre/diffs";
 import { CodeView, type CodeViewHandle } from "@pierre/diffs/react";
 import { useQuery, useSuspenseQueries, useSuspenseQuery } from "@tanstack/react-query";
@@ -109,12 +100,7 @@ import {
 	getChangesFileRowItems,
 	type FileRowItem,
 } from "./file-row.ts";
-import {
-	contiguousSelectionByLine,
-	contiguousSelectionsFromHunk,
-	rangeFromLineGroups,
-	synthesizeFilePatch,
-} from "#ui/hunk.ts";
+import { contiguousSelectionByLine } from "#ui/hunk.ts";
 import { buildIndexByKey, type NavigationIndex } from "#ui/workspace/navigation-index.ts";
 import { showNativeContextMenu, showNativeMenuFromTrigger } from "#ui/native-menu.ts";
 import { useFileMenuItems } from "#ui/routes/project/$id/workspace/useFileMenuItems.ts";
@@ -144,8 +130,13 @@ import {
 	useCommentCreate,
 } from "#ui/annotation.ts";
 import { FileIcon } from "#ui/components/FileIcon.tsx";
-
-type Annotation = { _tag: "local"; id: string };
+import {
+	type Annotation,
+	type DiffView,
+	type DiffViewFile,
+	getDiffView,
+	hunkOperandIdentityKey,
+} from "./diff-view.ts";
 
 export type DiffViewerHandle = CodeViewHandle<Annotation>;
 
@@ -160,9 +151,6 @@ const diffDefaults = {
 	diffOverflow: "scroll",
 	diffStyle: "split",
 } satisfies Partial<GUISettings>;
-
-const hunkOperandIdentityKey = (operand: HunkOperand): string =>
-	operandIdentityKey(hunkOperand(operand));
 
 const getCommitFileRowItems = ({
 	commitDetails,
@@ -196,140 +184,6 @@ const getCommitFileRowItems = ({
 				}),
 			),
 	];
-};
-
-const mkCodeViewItem = (
-	id: string,
-	change: TreeChange,
-	hunks: Array<DiffHunk>,
-): CodeViewDiffItem<Annotation> => {
-	const combinedFilePatch = synthesizeFilePatch(change, hunks);
-	const version = hash(combinedFilePatch);
-	const parsed = parsePatchFiles(combinedFilePatch, String(version));
-
-	const [patch, ...restPatches] = parsed;
-	if (!patch) throw new Error("Failed to parse any patches");
-	if (restPatches.length > 0) throw new Error("Parsed more than one patch");
-
-	const [fileDiff, ...restFiles] = patch.files;
-	if (!fileDiff) throw new Error("Failed to parse any files in patch");
-	if (restFiles.length > 0) throw new Error("Parsed more than one file in patch");
-
-	return {
-		type: "diff",
-		id,
-		version,
-		fileDiff,
-	};
-};
-
-type DiffViewDeps = {
-	fileParent: FileParent;
-	changes: Array<TreeChange>;
-	treeChangeDiffs: Array<UnifiedPatch | null>;
-};
-
-export type DiffViewFile = {
-	operand: FileOperand;
-	item: CodeViewDiffItem<Annotation>;
-	change: TreeChange;
-	patch: UnifiedPatch | null;
-	hunks: Array<DiffViewHunk>;
-};
-
-type DiffViewHunk = {
-	operand: HunkOperand;
-	selectedLines: CodeViewLineSelection;
-};
-
-type DiffView = {
-	navigationIndex: NavigationIndex<HunkOperand>;
-	items: Array<CodeViewDiffItem<Annotation>>;
-	fileByItemId: Map<string, DiffViewFile>;
-	fileByPath: Map<string, DiffViewFile>;
-	fileByHunkKey: Map<string, DiffViewFile>;
-	hunkByKey: Map<string, DiffViewHunk>;
-};
-
-/** Build relationships between our SDK data and Pierre's view. */
-const getDiffView = ({ fileParent, changes, treeChangeDiffs }: DiffViewDeps): DiffView => {
-	const navigationIndex: NavigationIndex<HunkOperand> = {
-		items: [],
-		indexByKey: new Map(),
-	};
-
-	const items: Array<CodeViewDiffItem<Annotation>> = [];
-
-	const fileByItemId = new Map<string, DiffViewFile>();
-	const fileByPath = new Map<string, DiffViewFile>();
-	const fileByHunkKey = new Map<string, DiffViewFile>();
-	const hunkByKey = new Map<string, DiffViewHunk>();
-
-	for (const [ci, change] of changes.entries()) {
-		const mdiff = treeChangeDiffs[ci];
-
-		const file: FileOperand = {
-			parent: fileParent,
-			path: change.path,
-		};
-		const item = mkCodeViewItem(
-			weakFileIdentityKey(file),
-			change,
-			mdiff && "subject" in mdiff && "hunks" in mdiff.subject ? mdiff.subject.hunks : [],
-		);
-
-		items.push(item);
-
-		const diffViewFile: DiffViewFile = {
-			operand: file,
-			item,
-			change,
-			patch: mdiff ?? null,
-			hunks: [],
-		};
-
-		fileByItemId.set(item.id, diffViewFile);
-		fileByPath.set(change.path, diffViewFile);
-
-		if (mdiff?.type === "Patch") {
-			for (const hunk of item.fileDiff.hunks) {
-				for (const selection of contiguousSelectionsFromHunk(hunk)) {
-					const range = rangeFromLineGroups(selection.lineGroups);
-					if (!range) continue;
-
-					const hunkOperand: HunkOperand = {
-						parent: file,
-						...selection,
-						isResultOfBinaryToTextConversion: mdiff.subject.isResultOfBinaryToTextConversion,
-					};
-					const hunkKey = hunkOperandIdentityKey(hunkOperand);
-
-					const len = navigationIndex.items.push(hunkOperand);
-					navigationIndex.indexByKey.set(hunkKey, len - 1);
-
-					const diffViewHunk: DiffViewHunk = {
-						operand: hunkOperand,
-						selectedLines: {
-							id: item.id,
-							range,
-						},
-					};
-					diffViewFile.hunks.push(diffViewHunk);
-					fileByHunkKey.set(hunkKey, diffViewFile);
-					hunkByKey.set(hunkKey, diffViewHunk);
-				}
-			}
-		}
-	}
-
-	return {
-		items,
-		fileByItemId,
-		fileByPath,
-		fileByHunkKey,
-		hunkByKey,
-		navigationIndex,
-	};
 };
 
 const withAnnotations = (
